Added architecture documents and initial implementation.
Architecture documents are contained in the folder architecture. Currently,
this folder contains the architecture documents for the CLE and the TF
framework.

The rest of the code is structured as follows: The folders nest and ros
contain the robot interfaces and implementations, whereas the folder nest
contains interface and implementation regarding the neuronal simulator. The
tf_framework folder contains the TF framework. Tests and mocks are contained
in a separate folder.

Patch3: Sorry, last version was a bit outdated. This one is more up to date
Patch4: Renamed packages, commented type annotations ('arrows')

Change-Id: Ib101af776dc4dbc4bd66c1113dafa4a279c08a4b