[hotfix] Fix segmentation fault in iCub tracking experiment.
This patch fixes a segmentation fault that appears to be caused
by an update to the OpenCV bridge. The issue would occur immediately
after pressing play on the iCub visual tracking experiment.

The root cause was code trying to convert a non RGB8 image using
an RGB8 conversion call. This is because the image generated by
the CV bridge conversion does not appear to default to RGB8 anymore
or the image provided by our gazebo_ros_plugin is not providing us
with an RGB8 image (less likely).

All other functions (such as red detection) had two features that
have been applied to the centroid detection function (where the
segfault was occuring):

- properly specifying the image conversion output format as RGB8

- checking for and ignoring NoneType input (which is provided by
  the transfer functions when the system first starts up and
  no images have been provided), this removes an error from the
  console that is not really an error:

    [ERROR]  'NoneType' object has no attribute 'encoding'

Change-Id: I2459de2cad9f305131c883e966c7bc4fb19d3fe6