[NRRPLT-4117] Consistently/properly use virtualgl on Lugano cluster.
This patch is part of the fix for the stuck on "0/0 assets loading", this patch
alone will not fix the issue.

When running Gazebo on the Lugano cluster, we were not always guaranteed to have
a valid 3D virtual X server because we were not properly loading the virtualgl
module. This patch aims to provide consistency to the virtual X environment so
that launches do not fail when they have improper 2D virtual X displays.

This patch includes:
- loading of the virtualgl module anytime we are running in a vgl context
- forcing of 32 bit colors (RGBA = 8 bits each including alpha vs default 24)
- forcing virtualgl calls through runtime redirects rather than failing to
  open an opengl library on each call and defaulting to the virtual call

Further debugging and performance enhancements are possible in the future by
adding the +v argument to vglrun and logging to a file, but these initial
parameters will provide consistency and stability to our launches.

Change-Id: If2e837cf0b5fc47ad1aba2de051cbbb9a33cd06b