[NRRPLT-4117] Use correct ROS packages on the Lugano cluster.
The Gazebo launch on Lugano servers is not using the same version
of the ROS packages as the cleserver. This was not explicitly an error,
but it is inconsistent.

In addition, the Jenkins plans for both ROS and the thirdparty packages
were updates to include fixes that are necessary to resolve the
"stuck on 0/0 bug", this patch allows for those fixes to actually be used
once they are deployed.

In the future, this code should be refactored to source the same script
as the cleserver - but for now this will allow for quick deployment and
testing to verify the 0/0 bug has been resolved without making other
changes that may introduce bugs.

Change-Id: I56852bfcefdc78de8efc4f9b51400cbb3f4770d3