[NRRPLT-4255] Do not expect output for Xvfb launch.
This patch resolves an issue where the first experiment launch on
a freshly started/rebooted cleserver would result in failure while
starting Xvfb.

The Xvfb launch does not produce any output when it is successfully
spawned for the first time. A logic error meant that this would
be interpreted as a failure to launch after a lengthy timeout.

Further, this means that the same Xvfb server that was spawned the first
time was used in each successive simulation as it would never be properly
killed. Another issue [NRRPLT-4256] will fix the cleanup process for
staled Xvfb/ssh processes in the case of launch failure.

Change-Id: I6070797bf760ea9e8ddc3549f4176430bf962104