[NRRPLT-4280] Fix Lugano Xvnc spawn conflict failure.
This patch fixes an issue where Xvnc would fail to launch on the
Lugano cluster because another Xvnc process was running on the same
port. Since we are not the only cluster users, we cannot guarantee
that all Xvnc display ports are available.

The previous code attempted to then find the display port of the
conflicting Xvnc process, but this code was malformed and failed
when multiple users were using Xvnc on the cluster node.

This patch replaces the random port selection with a systematic
iteration through the available ports 10-100, using the first valid
display port that is not in use. This avoids any conflicts with
other users. We cannot just ask Xvnc to select the next available
port because this does not appear to correctly launch on the RHEL
cluster nodes.

Change-Id: I63d33cc0b1cd0c8307872d1acd88bc91d9b576eb