[NRRPLT-4130] Only compile one version of numpy.
Previously, two versions of numpy were being build - 1.11.0 specified
in the requirements.txt and a newer version as a dependency of scipy.

This was because pip analyzes all dependencies and requirements when
given a set of packages to install by running their setup.py, in the
case of scipy, this would automatically kick off an easy_install of
numpy even though another version of numpy was specified to pip.

This patch forces numpy (and a Cython dependency) to be installed in
the virtualenv before any other packages, which resolves the dependency
ordering issue.

Change-Id: I654085705a44bf7908ec86fe4ebad279336ac470