[NRRPLT-4521] Add testfixtures to requirements.txt
This patch adds testfixtures to requirements.txt because it is used
by test files in the cle, but was not installed by 'make devinstall',
which caused 'make test' to fail on clean local installs.

testfixtures is included in the test extension requirements file, but
this is not installed by the devinstall. This is also the reason that
mock was included in both of the files.

Change-Id: I16db167c73344d963af9bfc0775ea8efc05e3d66