[NRRPLT-4750] Disconnect/disable TF I/O after delete.
This patch adds functionality to disconnect/disable a TF during
deletion rather than just disabling it from the update loop. This
is needed to prevent spike generating neurons from firing at the rates
they were set to before the TF was deleted. It is also not sufficient
to just reset the spike generator because the default user-supplied
rate my be above zero. The fix was extended to clean up robot publishers
and subscribers even though they did not directly cause an issue.

The patch includes fixes to both the robot and brain side:

- unregistering ROS topic subscriptions/publishers that are unused
- setting stimulus neurons to a zero rate and disabling them
- clearing the output buffers of integrators/records and disabling them

Test coverage is roughly the same percentage as before with minimal
additions.

Change-Id: I552e8999e7d9f0b17aa8b57e8a742311d16207cc