[NRRPLT-5091] Fix bad fmod precision with variable timestep.
This patch replaces a call to math.fmod with an explicit
definition to avoid floating point arithmetic errors encountered
when setting a custom simulation timestep.

e.g. If the simulation timestep was 0.999 (999 ms) and the simulator
timestep was 0.001 (1 ms), the previous arithmetic would abort
claiming that 0.999 cannot be cleanly divided by 0.001 (even though
it is clearly 999).

Change-Id: I3a4e3b17518088370171ff19430ad44a6fa57579