[NRRPLT-5284] Support deterministic Nest RNG seed.
This patch provides support for setting an experiment specific RNG
seed for Nest by overriding the pyNN setup function. This ensures
consistent behavior by overriding any setup calls within brain files.

Previously, the parameter to set the RNG seed was incorrect "grng_seeds"
instead of the valid "grng_seed" or "rng_seeds", this has been fixed.

It also has the benefit of ensuring we always set the number of threads
and number of RNG seeds consistently, regardless of what is in the
user's brain file. In the future, this can also be used to ensure
proper/consistent setup of the same RNG seed in multi-threaded simulations
(as the simulation should be deterministic regardless of the number of
  threads used).

Change-Id: I0df74ad04704d9fd81c2faf3d9fb326481f79b41