[NRRPLT-4853] Allow users to specify multiple PyNN/Nest threads.
This patch allows users to specify a number of threads in their
brain files via the standard pynn.Setup(...) mechanism. Previously
threads was always forced to 1 due to multi-threading issues that
have been addressed in the corresponding nest-simulator repo
patch.

The number of threads will default to 1 if unspecified and is capped
at the number of available cores - 1 to prevent oversubscribing, which
significantly impacts performance and since we cannot guarantee the
available target CPU configuration.

Change-Id: Id3ff7d624e7393a38a4e5c724798c6f08a8ed205